+++ examples/typegoose/entities/user.ts
@@ -5,7 +5,7 @@ import { Field, ObjectType } from "type-graphql";
 @ObjectType()
 export class User {
   @Field()
-  readonly id!: Types.ObjectId;
+  readonly _id!: Types.ObjectId;
 
   @Field()
   @Property({ required: true })