+++ examples/typegoose/helpers.ts
@@ -12,21 +12,21 @@ export async function seedDatabase() {
     {
       title: "Recipe 1",
       description: "Desc 1",
-      author: defaultUser.id,
+      author: defaultUser._id,
       ratings: [
-        { value: 2, user: defaultUser.id },
-        { value: 4, user: defaultUser.id },
-        { value: 5, user: defaultUser.id },
-        { value: 3, user: defaultUser.id },
-        { value: 4, user: defaultUser.id },
+        { value: 2, user: defaultUser._id },
+        { value: 4, user: defaultUser._id },
+        { value: 5, user: defaultUser._id },
+        { value: 3, user: defaultUser._id },
+        { value: 4, user: defaultUser._id },
       ],
     },
     {
       title: "Recipe 2",
-      author: defaultUser.id,
+      author: defaultUser._id,
       ratings: [
-        { value: 2, user: defaultUser },
-        { value: 4, user: defaultUser },
+        { value: 2, user: defaultUser._id },
+        { value: 4, user: defaultUser._id },
       ],
     },
   ]);